// src/WEBL/ClientApp/src/app/core/models/internalOrderItems.models.ts
export class internalOrderItems {
  stockId: number;
  originalValue: number;
  value: number;
  quantity: number;
  itemTotal: number;
  vatAppl: boolean;
  grnAppl: boolean;
  glcodeId: number;
  departmentId: number;
  uomid: number;
  packSize: number;
  totalUnits: number;
  total: number;
  departmentName: string
}